Modified DigitalDelivery so it inherits from Spree::ShippingCalculator. Fixes #48
